+++ src/components/spartan/SDTable/types.ts
@@ -1,7 +1,16 @@
 export type TDTableProps = {
     class?: string;
-    borderless: boolean;
+    borderless?: boolean;
     data: any[];
+    rowLink?: (row: any) => string;
+};
+
+export type TDColumnProps = {
+    field: string;
+    slots?: any;
+    header?: string;
+    noLink?: boolean;
+    sort?: 'asc' | 'des';
 };
 
 export type TDTableEmits = {};