+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Initial release of `SDTable.vue` with dynamic data display and customizable row content.
+- Column sorting with ascending and descending order for `SDTable.vue`.
+- Clickable rows with dynamic URL links for `SDTable.vue`.
+- Customizable headers with slots for icons and elements in `SDTable.vue`.
+- Borderless layout option for `SDTable.vue`.
 
 ## 2.0.0-beta.102 - 2024-09-17